Refactor SimServices component by moving it to a feature-based structure and extracting logic into a custom hook for improved organization and maintainability
