+++ src/features/sim-services/index.tsx
@@ -1,0 +1,28 @@
+import { useTranslation } from 'react-i18next';
+import S from './styled';
+import useSimServices from './hooks/use-sim-services';
+
+const SimServices = () => {
+  const { t } = useTranslation(['simservices']);
+  const { simNumbers, currentSimNumber, setCurrentSimNumber, play } =
+    useSimServices();
+
+  return (
+    <>
+      <S.MainContainer>
+        {simNumbers.map((x) => (
+          <S.ResultsBox onClick={() => setCurrentSimNumber(x)} key={x.id}>
+            <S.Item>{x.name}</S.Item>
+          </S.ResultsBox>
+        ))}
+      </S.MainContainer>
+      <S.ButtonContainer>
+        <button disabled={currentSimNumber === null} onClick={play}>
+          {t('play', { ns: 'global' })}
+        </button>
+      </S.ButtonContainer>
+    </>
+  );
+};
+
+export default SimServices;
